change type value

+++ src/main/java/org/example/shoppingcart/model/discount/PointDiscount.java
@@ -10,7 +10,7 @@ import static java.lang.Math.min;
 
 @Data
 public class PointDiscount extends Discount {
-    private Double value;
+    private int value;
     @Override
     public double applyDiscount(double total, List<Item> items) {
         if (value == 0.0) {
